pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "monaco"
version = "0.1.2"
description = "Task and project estimation using Monte Carlo simulation"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Martijn Wieriks"}
]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Scheduling",
    "Topic :: Scientific/Engineering :: Mathematics",
]
keywords = [
    "monte-carlo",
    "simulation",
    "estimation",
    "project-management",
    "task-planning",
    "forecasting",
]
dependencies = [
    "click>=7.0",
    "matplotlib>=3.0",
    "seaborn>=0.11",
    "numpy>=1.18",
    "PyYAML>=5.0",
    "networkx>=2.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.10",
    "mypy>=0.900",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
monaco = "monaco.cli:main"

[project.urls]
Homepage = "https://github.com/sepam/monaco"
Repository = "https://github.com/sepam/monaco"

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "matplotlib.*",
    "seaborn.*",
    "numpy.*",
    "click.*",
    "networkx.*",
    "yaml.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | \.ruff_cache
    | build
    | dist
)/
'''

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "SIM102", # nested if statements (can reduce readability for complex conditions)
]

[tool.ruff.lint.isort]
known-first-party = ["monaco"]

[tool.ruff.lint.per-file-ignores]
"src/monaco/__init__.py" = ["F403", "F401", "I001", "E402"]  # Star imports and order are intentional
